src/requirements.txt
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
et-xmlfile==1.1.0
h11==0.13.0
idna==3.3
joblib==1.1.0
lxml==4.9.0
nltk==3.7
openpyxl==3.0.10
outcome==1.1.0
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
python-magic==0.4.27
PyVirtualDisplay==3.0
regex==2022.6.2
requests==2.27.1
selenium==4.2.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
toml==0.10.2
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.7.0
wsproto==1.1.0
